+++ attalos/util/transformers/texttransformer.py
